Remove code climate coverage configuration
Code climate doesn't support windows at the moment:

Known Issue: Test Reporters Do Not Currently Support Windows
All of our test reporters do not currently work correctly when run from a Windows-based machine. Though our reporter won't throw an error when it runs, the coverage data that it POSTs to Code Climate will not be processed correctly. This is due to a bug in the server-side regular expression that we use to validate file paths within a payload. We have a fix planned for this issue.